fix(stm32): Issue #21 요청 반영
변경점)
 - AngleMessageQueue 요소 타입 변경
    uint32_t -> uint16_t
 - MainTaskHandler 이름 변경

+++ App/Interrupt/ISR_callback.cpp
@@ -7,7 +7,7 @@ extern osMessageQueueId_t AngleQueueHandle;
 
 extern "C"
 void HAL_UARTEx_RxEventCallback(UART_HandleTypeDef *huart, uint16_t Size){
-  uint32_t angle {0};
+  uint16_t angle {0};
